+++ src/admin/contexts/courseContext/CourseProvider.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useState, type ReactNode } from "react";
+import { useCallback, useMemo, useState, type ReactNode } from "react";
 import { CourseContext } from "./CourseContext";
 import type { CourseContextType } from "./CourseContext.type";
 import { CourseService } from "../../services/course/CourseService";
@@ -6,12 +6,14 @@ import type {
   CreateCoursePayload,
   CourseResponseDto,
   UpdateCoursePayload,
-} from "../../services/course/CourseService";
-import type {
-  GetPaginated,
-  PaginatedData,
-} from "../../../shared/types/PaginacionType";
-import { useHandleApiError } from "../../../shared/hooks/useHandleApiError";
+} from "@/admin/types/course.types";
+import {
+  useHandleApiError,
+  type GetPaginated,
+  type PaginatedData,
+  useToaster,
+  withLoading,
+} from "@/shared";
 
 interface CourseProviderProps {
   children: ReactNode;
@@ -19,7 +21,7 @@ interface CourseProviderProps {
 
 export const CourseProvider: React.FC<CourseProviderProps> = ({ children }) => {
   const { handleApiError } = useHandleApiError();
-
+  const { showToast } = useToaster();
   const [loading, setLoading] = useState<boolean>(false);
   const [courses, setCourses] = useState<CourseResponseDto[]>([]);
   const [paginatedCourse, setPaginatedCourse] =
@@ -27,92 +29,137 @@ export const CourseProvider: React.FC<CourseProviderProps> = ({ children }) => {
   const [selectedCourse, setSelectedCourse] =
     useState<CourseResponseDto | null>(null);
 
-  const registerCourse = async (data: CreateCoursePayload): Promise<void> => {
-    setLoading(true);
-    try {
-      await CourseService.registerCourseApi(data);
-    } catch (error) {
-      handleApiError(error, "Error al crear el curso");
-    } finally {
-      setLoading(false);
-    }
-  };
+  const registerCourse = useCallback(
+    async (data: CreateCoursePayload): Promise<void> => {
+      await withLoading(async () => {
+        try {
+          await CourseService.registerCourseApi(data);
 
-  const updateCourse = async (data: UpdateCoursePayload): Promise<void> => {
-    setLoading(true);
-    try {
-      await CourseService.updateCourseApi(data);
-    } catch (error) {
-      handleApiError(error, "Error al actualizar el curso");
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const getCourse = useCallback(async () => {
-    setLoading(true);
-    try {
-      const response = await CourseService.getCourseApi();
-      setCourses(response.data.data);
-    } catch (error) {
-      handleApiError(error, "Error al obtener los cursos");
-    } finally {
-      setLoading(false);
-    }
-  }, []);
-
-  const getCourseById = async (
-    id: number
-  ): Promise<CourseResponseDto | undefined> => {
-    setLoading(true);
-    try {
-      const courseData = await CourseService.getCourseByIdApi(id);
-      return courseData;
-    } catch (error) {
-      handleApiError(error, "Error al obtener el curso");
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const getPaginatedCourse = useCallback(
-    async (params: GetPaginated): Promise<void> => {
-      setLoading(true);
-      try {
-        const response = await CourseService.getPaginatedCourseApi(params);
-        setPaginatedCourse(response.data);
-      } catch (error) {
-        handleApiError(error, "Error al obtener los cursos paginados");
-      } finally {
-        setLoading(false);
-      }
+          showToast({
+            title: "Curso creado exitosamente",
+            message: "El curso ha sido creado exitosamente",
+            type: "success",
+            position: "bottom-right",
+          });
+        } catch (error) {
+          handleApiError(error, "Error al crear el curso");
+          throw error;
+        }
+      }, setLoading);
     },
     []
   );
 
-  const deleteCourse = async (id: number): Promise<void> => {
-    setLoading(true);
-    try {
-      await CourseService.deleteCourseApi(id);
-    } catch (error) {
-      handleApiError(error, "Error al eliminar el curso");
-    } finally {
-      setLoading(false);
-    }
-  };
-  const contextValue: CourseContextType = {
-    loading,
-    courses,
-    paginatedCourse,
-    selectedCourse,
-    registerCourse,
-    updateCourse,
-    getCourse,
-    getCourseById,
-    getPaginatedCourse,
-    deleteCourse,
-    setSelectedCourse,
-  };
+  const updateCourse = useCallback(
+    async (data: UpdateCoursePayload): Promise<void> => {
+      await withLoading(async () => {
+        try {
+          await CourseService.updateCourseApi(data);
+
+          showToast({
+            title: "Curso actualizado exitosamente",
+            message: "El curso ha sido actualizado exitosamente",
+            type: "success",
+            position: "bottom-right",
+          });
+        } catch (error) {
+          handleApiError(error, "Error al actualizar el curso");
+          throw error;
+        }
+      }, setLoading);
+    },
+    []
+  );
+
+  const getCourse = useCallback(async (): Promise<void> => {
+    await withLoading(async () => {
+      try {
+        const response = await CourseService.getCourseApi();
+        setCourses(response.data.data);
+      } catch (error) {
+        handleApiError(error, "Error al obtener los cursos");
+        throw error;
+      }
+    }, setLoading);
+  }, []);
+
+  const getCourseById = useCallback(
+    async (id: number): Promise<CourseResponseDto | undefined> => {
+      return await withLoading(async () => {
+        try {
+          return await CourseService.getCourseByIdApi(id);
+        } catch (error) {
+          handleApiError(error, "Error al obtener el curso");
+          throw error;
+        }
+      }, setLoading);
+    },
+    []
+  );
+
+  const getPaginatedCourse = useCallback(
+    async (params: GetPaginated): Promise<void> => {
+      await withLoading(async () => {
+        try {
+          const response = await CourseService.getPaginatedCourseApi(params);
+          setPaginatedCourse(response.data);
+        } catch (error) {
+          handleApiError(error, "Error al obtener cursos paginados");
+          throw error;
+        }
+      }, setLoading);
+    },
+    []
+  );
+
+  const deleteCourse = useCallback(async (id: number): Promise<void> => {
+    await withLoading(async () => {
+      try {
+        await CourseService.deleteCourseApi(id);
+      } catch (error) {
+        handleApiError(error, "Error al eliminar el curso");
+        throw error;
+      }
+    }, setLoading);
+  }, []);
+
+  const actions = useMemo(
+    () => ({
+      registerCourse,
+      updateCourse,
+      getCourse,
+      getCourseById,
+      getPaginatedCourse,
+      deleteCourse,
+      setSelectedCourse,
+    }),
+    [
+      registerCourse,
+      updateCourse,
+      getCourse,
+      getCourseById,
+      getPaginatedCourse,
+      deleteCourse,
+    ]
+  );
+
+  const state = useMemo(
+    () => ({
+      loading,
+      courses,
+      paginatedCourse,
+      selectedCourse,
+    }),
+    [loading, courses, paginatedCourse, selectedCourse]
+  );
+
+  const contextValue: CourseContextType = useMemo(
+    () => ({
+      ...state,
+      ...actions,
+    }),
+    [state, actions]
+  );
 
   return (
     <CourseContext.Provider value={contextValue}>
