+++ src/admin/hooks/hooksUI/Year/useCreateYearView.ts
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import { useNavigate, useParams } from "react-router-dom";
-import { useYear } from "../useYear";
+import { useYear } from "@/admin";
 import { useHandleApiError } from "@/shared";
 
 interface FormData {
