+++ src/admin/hooks/hooksUI/index.ts
@@ -1,0 +1,4 @@
+export { useCourseView } from "./Course/useCourseView";
+export { useCreateCourseView } from "./Course/useCreateCourseView";
+export { useCreateYearView } from "./Year/useCreateYearView";
+export { useListYear } from "./Year/useListYear";
