+++ src/admin/hooks/index.ts
@@ -1,0 +1,7 @@
+export { useCourse } from "./useCourse";
+export { useStatistics } from "./useStatistics";
+export { useStudent } from "./useStudent";
+export { useSubject } from "./useSubject";
+export { useTeacher } from "./useTeacher";
+export { useYear } from "./useYear";
+export * from "./hooksUI";
