+++ src/admin/types/index.ts
@@ -1,0 +1,13 @@
+export type {
+  CourseResponseDto,
+  CreateCoursePayload,
+  PaginatedCourseResponse,
+  UpdateCoursePayload,
+} from "./course.types";
+export type { StatisticsResponseDto } from "./statistics.types";
+export type {
+  CreateYearPayload,
+  PaginatedYearResponse,
+  UpdateYearPayload,
+  YearResponseDto,
+} from "./year.types";
